@@ -4,14 +4,15 @@ description: Simple yet powerful, open-source SpaceX launch tracker.
 version: 2.14.0+80
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   add_2_calendar: ^2.1.0
-  big_tip: ^2.0.0
+  # big_tip: ^2.0.0
+  big_tip: ^3.0.0
   cached_network_image: ^3.1.0
   cherry_components: ^0.3.1
   decorated_icon: ^1.2.1
