@@ -7,19 +7,19 @@ import 'index.dart';
 
 /// Specific details about an one-of-a-kink space payload.
 class Payload extends Equatable {
-  final CapsuleDetails capsule;
-  final String name;
-  final bool reused;
-  final String customer;
-  final String nationality;
-  final String manufacturer;
-  final num mass;
-  final String orbit;
-  final num periapsis;
-  final num apoapsis;
-  final num inclination;
-  final num period;
-  final String id;
+  final CapsuleDetails? capsule;
+  final String? name;
+  final bool? reused;
+  final String? customer;
+  final String? nationality;
+  final String? manufacturer;
+  final num? mass;
+  final String? orbit;
+  final num? periapsis;
+  final num? apoapsis;
+  final num? inclination;
+  final num? period;
+  final String? id;
 
   const Payload({
     this.capsule,
@@ -84,19 +84,19 @@ class Payload extends Equatable {
 
   String getPeriapsis(BuildContext context) => periapsis == null
       ? context.translate('spacex.other.unknown')
-      : '${NumberFormat.decimalPattern().format(periapsis.round())} km';
+      : '${NumberFormat.decimalPattern().format(periapsis!.round())} km';
 
   String getApoapsis(BuildContext context) => apoapsis == null
       ? context.translate('spacex.other.unknown')
-      : '${NumberFormat.decimalPattern().format(apoapsis.round())} km';
+      : '${NumberFormat.decimalPattern().format(apoapsis!.round())} km';
 
   String getInclination(BuildContext context) => inclination == null
       ? context.translate('spacex.other.unknown')
-      : '${NumberFormat.decimalPattern().format(inclination.round())}°';
+      : '${NumberFormat.decimalPattern().format(inclination!.round())}°';
 
   String getPeriod(BuildContext context) => period == null
       ? context.translate('spacex.other.unknown')
-      : '${NumberFormat.decimalPattern().format(period.round())} min';
+      : '${NumberFormat.decimalPattern().format(period!.round())} min';
 
   bool get isNasaPayload =>
       customer == 'NASA (CCtCap)' ||
@@ -105,18 +105,18 @@ class Payload extends Equatable {
 
   @override
   List<Object> get props => [
-        capsule,
-        name,
-        reused,
-        customer,
-        nationality,
-        manufacturer,
-        mass,
-        orbit,
-        periapsis,
-        apoapsis,
-        inclination,
-        period,
-        id,
+        capsule!,
+        name!,
+        reused!,
+        customer!,
+        nationality!,
+        manufacturer!,
+        mass!,
+        orbit!,
+        periapsis!,
+        apoapsis!,
+        inclination!,
+        period!,
+        id!,
       ];
 }
