@@ -15,9 +15,9 @@ class CorePage extends StatelessWidget {
   final String coreId;
 
   const CorePage({
-    Key key,
-    this.launchId,
-    this.coreId,
+    Key? key,
+    required this.launchId,
+    required this.coreId,
   }) : super(key: key);
 
   static const route = '/core';
@@ -27,64 +27,68 @@ class CorePage extends StatelessWidget {
     final core = context
         .watch<LaunchesCubit>()
         .getLaunch(launchId)
-        .rocket
-        .getCore(coreId);
+        ?.rocket
+        ?.getCore(coreId);
 
     return Scaffold(
       body: SliverPage(
         title: context.translate(
           'spacex.dialog.vehicle.title_core',
-          parameters: {'serial': core.serial},
+          parameters: {'serial': core!.serial!},
         ),
         header: SwiperHeader(list: List.from(SpaceXPhotos.cores)..shuffle()),
         children: <Widget>[
           SliverSafeArea(
             top: false,
             sliver: SliverToBoxAdapter(
-              child: RowLayout.body(children: <Widget>[
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.model'),
-                  core.getBlock(context),
-                ),
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.status'),
-                  core.getStatus,
-                ),
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.first_launched'),
-                  core.getFirstLaunched(context),
-                ),
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.launches'),
-                  core.getLaunches,
-                ),
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.landings_rtls'),
-                  core.getRtlsLandings,
-                ),
-                RowItem.text(
-                  context.translate('spacex.dialog.vehicle.landings_asds'),
-                  core.getAsdsLandings,
-                ),
-                Separator.divider(),
-                if (core.hasMissions) ...[
-                  for (final mission in core.launches)
-                    RowTap(
-                      context.translate(
-                        'spacex.dialog.vehicle.mission',
-                        parameters: {'number': mission.flightNumber.toString()},
+              child: RowLayout.body(
+                children: <Widget>[
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.model'),
+                    core.getBlock(context),
+                  ),
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.status'),
+                    core.getStatus!,
+                  ),
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.first_launched'),
+                    core.getFirstLaunched(context),
+                  ),
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.launches'),
+                    core.getLaunches,
+                  ),
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.landings_rtls'),
+                    core.getRtlsLandings,
+                  ),
+                  RowItem.text(
+                    context.translate('spacex.dialog.vehicle.landings_asds'),
+                    core.getAsdsLandings,
+                  ),
+                  Separator.divider(),
+                  if (core.hasMissions) ...[
+                    for (final mission in core.launches!)
+                      RowTap(
+                        context.translate(
+                          'spacex.dialog.vehicle.mission',
+                          parameters: {
+                            'number': mission.flightNumber.toString()
+                          },
+                        ),
+                        mission.name,
+                        onTap: () => Navigator.pushNamed(
+                          context,
+                          LaunchPage.route,
+                          arguments: {'id': mission.id},
+                        ),
                       ),
-                      mission.name,
-                      onTap: () => Navigator.pushNamed(
-                        context,
-                        LaunchPage.route,
-                        arguments: {'id': mission.id},
-                      ),
-                    ),
-                  Separator.divider()
+                    Separator.divider()
+                  ],
+                  ExpandText(core.getDetails(context))
                 ],
-                ExpandText(core.getDetails(context))
-              ]),
+              ),
             ),
           ),
         ],
