@@ -19,30 +19,30 @@ class HomeTab extends StatefulWidget {
 }
 
 class _HomeTabState extends State<HomeTab> {
-  ScrollController _controller;
+  ScrollController? _controller;
   double _offset = 0.0;
 
   @override
   void initState() {
     super.initState();
     _controller = ScrollController()
-      ..addListener(() => setState(() => _offset = _controller.offset));
+      ..addListener(() => setState(() => _offset = _controller!.offset));
   }
 
   @override
   Widget build(BuildContext context) {
     return RequestSliverPage<LaunchesCubit, List<List<Launch>>>(
-      controller: _controller,
+      controller: _controller!,
       title: context.translate('spacex.home.title'),
       popupMenu: Menu.home,
       headerBuilder: (context, state, value) => _HeaderView(
-        launch: LaunchUtils.getUpcomingLaunch(value),
+        launch: LaunchUtils.getUpcomingLaunch(value!)!,
         offset: _offset,
       ),
       childrenBuilder: (context, state, value) => <Widget>[
         SliverToBoxAdapter(
           child: _HomeView(
-            LaunchUtils.getUpcomingLaunch(value),
+            LaunchUtils.getUpcomingLaunch(value)!,
           ),
         ),
       ],
@@ -55,9 +55,9 @@ class _HeaderView extends StatelessWidget {
   final double offset;
 
   const _HeaderView({
-    Key key,
-    this.launch,
-    this.offset,
+    Key? key,
+    required this.launch,
+    required this.offset,
   }) : super(key: key);
 
   @override
@@ -127,7 +127,7 @@ class _HeaderView extends StatelessWidget {
 class _HomeView extends StatelessWidget {
   final Launch launch;
 
-  const _HomeView(this.launch, {Key key}) : super(key: key);
+  const _HomeView(this.launch, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -136,9 +136,9 @@ class _HomeView extends StatelessWidget {
         icon: Icons.public,
         title: context.translate(
           'spacex.home.tab.mission.title',
-          parameters: {'rocket': launch.rocket.name},
+          parameters: {'rocket': launch.rocket!.name!},
         ),
-        subtitle: payloadSubtitle(context, launch.rocket.payloads),
+        subtitle: payloadSubtitle(context, launch.rocket!.payloads!),
         onTap: () => Navigator.pushNamed(
           context,
           LaunchPage.route,
@@ -166,10 +166,10 @@ class _HomeView extends StatelessWidget {
         onTap: !launch.tentativeTime
             ? () async {
                 await Add2Calendar.addEvent2Cal(Event(
-                  title: launch.name,
+                  title: launch.name!,
                   description: launch.details ??
                       context.translate('spacex.launch.page.no_description'),
-                  location: launch.launchpad.name ??
+                  location: launch.launchpad?.name ??
                       context.translate('spacex.other.unknown'),
                   startDate: launch.localLaunchDate,
                   endDate: launch.localLaunchDate.add(
@@ -185,7 +185,7 @@ class _HomeView extends StatelessWidget {
         title: context.translate('spacex.home.tab.launchpad.title'),
         subtitle: context.translate(
           'spacex.home.tab.launchpad.body',
-          parameters: {'launchpad': launch.launchpad.name},
+          parameters: {'launchpad': launch.launchpad!.name!},
         ),
         onTap: launch.launchpad != null
             ? () => Navigator.pushNamed(
@@ -202,18 +202,18 @@ class _HomeView extends StatelessWidget {
         subtitle: launch.staticFireDate == null
             ? context.translate('spacex.home.tab.static_fire.body_unknown')
             : context.translate(
-                launch.staticFireDate.isBefore(DateTime.now())
+                launch.staticFireDate!.isBefore(DateTime.now())
                     ? 'spacex.home.tab.static_fire.body_done'
                     : 'spacex.home.tab.static_fire.body',
                 parameters: {'date': launch.getStaticFireDate(context)},
               ),
       ),
       Separator.divider(indent: 72),
-      if (launch.rocket.hasFairings)
+      if (launch.rocket!.hasFairings!)
         ListCell.icon(
           icon: Icons.directions_boat,
           title: context.translate('spacex.home.tab.fairings.title'),
-          subtitle: fairingSubtitle(context, launch.rocket.fairings),
+          subtitle: fairingSubtitle(context, launch.rocket!.fairings!),
         )
       else
         ListCell(
@@ -227,8 +227,8 @@ class _HomeView extends StatelessWidget {
                 : null,
           ),
           title: context.translate('spacex.home.tab.capsule.title'),
-          subtitle: capsuleSubtitle(context, launch.rocket.getSinglePayload),
-          onTap: launch.rocket.hasCapsule
+          subtitle: capsuleSubtitle(context, launch.rocket!.getSinglePayload),
+          onTap: launch.rocket!.hasCapsule
               ? () => Navigator.pushNamed(
                     context,
                     CapsulePage.route,
@@ -238,7 +238,7 @@ class _HomeView extends StatelessWidget {
         ),
       Separator.divider(indent: 72),
       AbsorbPointer(
-        absorbing: launch.rocket.isFirstStageNull,
+        absorbing: launch.rocket!.isFirstStageNull,
         child: ListCell(
           leading: SvgPicture.asset(
             'assets/icons/fins.svg',
@@ -250,25 +250,25 @@ class _HomeView extends StatelessWidget {
                 : null,
           ),
           title: context.translate('spacex.home.tab.first_stage.title'),
-          subtitle: launch.rocket.isHeavy
+          subtitle: launch.rocket!.isHeavy
               ? context.translate(
-                  launch.rocket.isFirstStageNull
+                  launch.rocket!.isFirstStageNull
                       ? 'spacex.home.tab.first_stage.body_null'
                       : 'spacex.home.tab.first_stage.heavy_dialog.body',
                 )
               : coreSubtitle(
                   context: context,
-                  core: launch.rocket.getSingleCore,
+                  core: launch.rocket!.getSingleCore,
                   isSideCore:
-                      launch.rocket.isSideCore(launch.rocket.getSingleCore),
+                      launch.rocket!.isSideCore(launch.rocket!.getSingleCore),
                 ),
-          onTap: !launch.rocket.isFirstStageNull
-              ? () => launch.rocket.isHeavy
+          onTap: !launch.rocket!.isFirstStageNull
+              ? () => launch.rocket!.isHeavy
                   ? showHeavyDialog(context, launch)
                   : openCorePage(
                       context: context,
-                      launchId: launch.id,
-                      coreId: launch.rocket.getSingleCore.id,
+                      launchId: launch.id!,
+                      coreId: launch.rocket!.getSingleCore.id!,
                     )
               : null,
         ),
@@ -277,14 +277,14 @@ class _HomeView extends StatelessWidget {
       ListCell.icon(
         icon: Icons.center_focus_weak,
         title: context.translate('spacex.home.tab.landing.title'),
-        subtitle: landingSubtitle(context, launch.rocket.getSingleCore),
-        onTap: launch.rocket.getSingleCore.landpad != null
+        subtitle: landingSubtitle(context, launch.rocket!.getSingleCore),
+        onTap: launch.rocket!.getSingleCore.landpad != null
             ? () => Navigator.pushNamed(
                   context,
                   LandpadPage.route,
                   arguments: {
                     'launchId': launch.id,
-                    'coreId': launch.rocket.getSingleCore.id,
+                    'coreId': launch.rocket!.getSingleCore.id,
                   },
                 )
             : null,
@@ -293,7 +293,7 @@ class _HomeView extends StatelessWidget {
     ]);
   }
 
-  void openCorePage({BuildContext context, String launchId, String coreId}) {
+  void openCorePage({required BuildContext context, required String launchId, required String coreId}) {
     Navigator.pushNamed(
       context,
       CorePage.route,
@@ -312,14 +312,14 @@ class _HomeView extends StatelessWidget {
         ),
         padding: EdgeInsets.zero,
         children: [
-          for (final core in upcomingLaunch.rocket.cores)
+          for (final core in upcomingLaunch.rocket!.cores!)
             AbsorbPointer(
               absorbing: core.id == null,
               child: ListCell(
                 title: core.id != null
                     ? context.translate(
                         'spacex.dialog.vehicle.title_core',
-                        parameters: {'serial': core.serial},
+                        parameters: {'serial': core.serial!},
                       )
                     : context.translate(
                         'spacex.home.tab.first_stage.heavy_dialog.core_null_title',
@@ -327,12 +327,12 @@ class _HomeView extends StatelessWidget {
                 subtitle: coreSubtitle(
                   context: context,
                   core: core,
-                  isSideCore: upcomingLaunch.rocket.isSideCore(core),
+                  isSideCore: upcomingLaunch.rocket!.isSideCore(core),
                 ),
                 onTap: () => openCorePage(
                   context: context,
-                  launchId: upcomingLaunch.id,
-                  coreId: core.id,
+                  launchId: upcomingLaunch.id!,
+                  coreId: core.id!,
                 ),
                 contentPadding: EdgeInsets.symmetric(
                   horizontal: 20,
@@ -346,17 +346,17 @@ class _HomeView extends StatelessWidget {
   String landingSubtitle(BuildContext context, Core core) {
     if (core.landingAttempt == null) {
       return context.translate('spacex.home.tab.landing.body_null');
-    } else if (!core.landingAttempt) {
+    } else if (!core.landingAttempt!) {
       return context.translate('spacex.home.tab.landing.body_expended');
     } else if (core.landpad == null && core.landingType != null) {
       return context.translate(
         'spacex.home.tab.landing.body_type',
-        parameters: {'type': core.landingType},
+        parameters: {'type': core.landingType!},
       );
     } else {
       return context.translate(
         'spacex.home.tab.landing.body',
-        parameters: {'zone': core.landpad.name},
+        parameters: {'zone': core.landpad!.name!},
       );
     }
   }
@@ -374,8 +374,8 @@ class _HomeView extends StatelessWidget {
         context.translate(
               'spacex.home.tab.mission.body_payload',
               parameters: {
-                'name': payloads[i].name,
-                'orbit': payloads[i].orbit
+                'name': payloads[i].name!,
+                'orbit': payloads[i].orbit!
               },
             ) +
             (i + 1 == payloadList.length ? '' : ', '),
@@ -413,7 +413,7 @@ class _HomeView extends StatelessWidget {
                   },
                 );
 
-  String coreSubtitle({BuildContext context, Core core, bool isSideCore}) =>
+  String coreSubtitle({required BuildContext context, required Core core, required bool isSideCore}) =>
       core.id == null || core.reused == null
           ? context.translate('spacex.home.tab.first_stage.body_null')
           : context.translate(
@@ -425,7 +425,7 @@ class _HomeView extends StatelessWidget {
                       : 'spacex.home.tab.first_stage.booster',
                 ),
                 'reused': context.translate(
-                  core.reused
+                  core.reused!
                       ? 'spacex.home.tab.first_stage.body_reused'
                       : 'spacex.home.tab.first_stage.body_new',
                 ),
@@ -438,7 +438,7 @@ class _HomeView extends StatelessWidget {
           : context.translate(
               'spacex.home.tab.capsule.body',
               parameters: {
-                'reused': payload.reused
+                'reused': payload.reused!
                     ? context.translate(
                         'spacex.home.tab.capsule.body_reused',
                       )
