@@ -18,12 +18,13 @@ typedef RequestListBuilderLoaded<T> = List<Widget> Function(
 /// Used when the desired page doesn't have slivers or reloading.
 class SimplePage extends StatelessWidget {
   final String title;
-  final Widget body, fab;
-  final List<Widget> actions;
+  final Widget body;
+  final Widget? fab;
+  final List<Widget>? actions;
 
   const SimplePage({
-    @required this.title,
-    @required this.body,
+    required this.title,
+    required this.body,
     this.fab,
     this.actions,
   });
@@ -51,13 +52,13 @@ class SimplePage extends StatelessWidget {
 /// This page also has state control via a `RequestCubit` parameter.
 class RequestSimplePage<C extends RequestCubit, T> extends StatelessWidget {
   final String title;
-  final Widget fab;
+  final Widget? fab;
   final RequestWidgetBuilderLoaded<T> childBuilder;
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   const RequestSimplePage({
-    @required this.title,
-    @required this.childBuilder,
+    required this.title,
+    required this.childBuilder,
     this.fab,
     this.actions,
   });
@@ -86,13 +87,13 @@ class RequestSimplePage<C extends RequestCubit, T> extends StatelessWidget {
 class SliverPage extends StatelessWidget {
   final String title;
   final Widget header;
-  final List<Widget> children, actions;
-  final Map<String, String> popupMenu;
-  final ScrollController controller;
+  final List<Widget>? children, actions;
+  final Map<String, String>? popupMenu;
+  final ScrollController? controller;
 
   const SliverPage({
-    @required this.title,
-    @required this.header,
+    required this.title,
+    required this.header,
     this.children,
     this.actions,
     this.popupMenu,
@@ -112,20 +113,20 @@ class SliverPage extends StatelessWidget {
             if (popupMenu != null)
               PopupMenuButton<String>(
                 itemBuilder: (context) => [
-                  for (final item in popupMenu.keys)
+                  for (final item in popupMenu!.keys)
                     PopupMenuItem(
                       value: item,
                       child: Text(FlutterI18n.translate(context, item)),
                     )
                 ],
                 onSelected: (text) =>
-                    Navigator.pushNamed(context, popupMenu[text]),
+                    Navigator.pushNamed(context, popupMenu![text]!),
                 icon: IconShadow(Icons.adaptive.more),
               ),
-            if (actions != null) ...actions,
+            if (actions != null) ...actions!,
           ],
         ),
-        ...children,
+        ...children!,
       ],
     );
   }
@@ -139,14 +140,14 @@ class RequestSliverPage<C extends RequestCubit, T> extends StatelessWidget {
   final String title;
   final RequestWidgetBuilderLoaded<T> headerBuilder;
   final RequestListBuilderLoaded<T> childrenBuilder;
-  final List<Widget> actions;
-  final Map<String, String> popupMenu;
-  final ScrollController controller;
+  final List<Widget>? actions;
+  final Map<String, String>? popupMenu;
+  final ScrollController? controller;
 
   const RequestSliverPage({
-    @required this.title,
-    @required this.headerBuilder,
-    @required this.childrenBuilder,
+    required this.title,
+    required this.headerBuilder,
+    required this.childrenBuilder,
     this.controller,
     this.actions,
     this.popupMenu,
@@ -158,11 +159,11 @@ class RequestSliverPage<C extends RequestCubit, T> extends StatelessWidget {
       onRefresh: () => context.read<C>().loadData(),
       child: RequestBuilder<C, T>(
         onInit: (context, state) => SliverPage(
-          controller: controller,
+          controller: controller!,
           title: title,
           header: Separator.none(),
-          actions: actions,
-          popupMenu: popupMenu,
+          actions: actions!,
+          popupMenu: popupMenu!,
         ),
         onLoading: (context, state, value) => SliverPage(
           controller: controller,
@@ -182,7 +183,7 @@ class RequestSliverPage<C extends RequestCubit, T> extends StatelessWidget {
           header: headerBuilder(context, state, value),
           actions: actions,
           popupMenu: popupMenu,
-          children: childrenBuilder(context, state, value),
+          children: childrenBuilder(context, state, value!),
         ),
         onError: (context, state, error) => SliverPage(
           controller: controller,
