@@ -30,15 +30,15 @@ Future<T> showPatreonDialog<T>(BuildContext context) {
           Text(
             context.translate('about.patreon.body_dialog'),
             textAlign: TextAlign.justify,
-            style: Theme.of(context).textTheme.subtitle1.copyWith(
-                  color: Theme.of(context).textTheme.caption.color,
+            style: Theme.of(context).textTheme.subtitle1?.copyWith(
+                  color: Theme.of(context).textTheme.caption?.color,
                 ),
           ),
           for (String patreon in _patreons)
             Text(
               patreon,
-              style: Theme.of(context).textTheme.subtitle1.copyWith(
-                    color: Theme.of(context).textTheme.caption.color,
+              style: Theme.of(context).textTheme.subtitle1?.copyWith(
+                    color: Theme.of(context).textTheme.caption?.color,
                   ),
             ),
           if (Theme.of(context).platform != TargetPlatform.iOS)
@@ -47,12 +47,12 @@ Future<T> showPatreonDialog<T>(BuildContext context) {
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
-                  FlatButton(
+                  TextButton(
                     onPressed: () => Navigator.pop(context, false),
                     child: Text(
                       context.translate('about.patreon.dismiss'),
-                      style: Theme.of(context).textTheme.bodyText2.copyWith(
-                            color: Theme.of(context).textTheme.caption.color,
+                      style: Theme.of(context).textTheme.bodyText2?.copyWith(
+                            color: Theme.of(context).textTheme.caption?.color,
                           ),
                     ),
                   ),
