@@ -8,7 +8,7 @@ import '../../utils/index.dart';
 /// Widget that tells the user that there's been an error in a network process.
 /// It allows the user to perform a reload action.
 class ErrorView<C extends RequestCubit> extends StatelessWidget {
-  const ErrorView({Key key}) : super(key: key);
+  const ErrorView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -17,14 +17,16 @@ class ErrorView<C extends RequestCubit> extends StatelessWidget {
         context.translate('spacex.other.loading_error.message'),
         style: Theme.of(context).textTheme.subtitle1,
       ),
-      action: Text(
-        context.translate('spacex.other.loading_error.reload'),
-        style: Theme.of(context).textTheme.subtitle1.copyWith(
-              color: Theme.of(context).accentColor,
-              fontWeight: FontWeight.bold,
-            ),
+      action: TextButton(
+        child: Text(
+          context.translate('spacex.other.loading_error.reload'),
+          style: Theme.of(context).textTheme.subtitle1?.copyWith(
+                color: Theme.of(context).accentColor,
+                fontWeight: FontWeight.bold,
+              ),
+        ),
+        onPressed: () => context.read<C>().loadData(),
       ),
-      actionCallback: () => context.read<C>().loadData(),
       child: Icon(Icons.cloud_off),
     );
   }
@@ -35,7 +37,7 @@ class ErrorView<C extends RequestCubit> extends StatelessWidget {
 /// Tells the user that there's been an error in a network process.
 /// It allows the user to perform a reload action.
 class ErrorSliverView<C extends RequestCubit> extends StatelessWidget {
-  const ErrorSliverView({Key key}) : super(key: key);
+  const ErrorSliverView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
