@@ -10,16 +10,16 @@ class MockVehiclesRepository extends Mock implements VehiclesRepository {}
 
 void main() {
   group('VehiclesCubit', () {
-    VehiclesCubit cubit;
-    MockVehiclesRepository repository;
+    VehiclesCubit? cubit;
+    MockVehiclesRepository repository = MockVehiclesRepository();
 
     setUp(() {
-      repository = MockVehiclesRepository();
+      // repository = MockVehiclesRepository();
       cubit = VehiclesCubit(repository);
     });
 
     tearDown(() {
-      cubit.close();
+      cubit!.close();
     });
 
     group('fetchData', () {
@@ -29,7 +29,7 @@ void main() {
           when(repository.fetchData()).thenAnswer(
             (_) => Future.value(const [RocketVehicle(id: '1')]),
           );
-          return cubit;
+          return cubit!;
         },
         act: (cubit) async => cubit.loadData(),
         verify: (_) => verify(repository.fetchData()).called(2),
@@ -43,7 +43,7 @@ void main() {
         'can throw an exception',
         build: () {
           when(repository.fetchData()).thenThrow(Exception('wtf'));
-          return cubit;
+          return cubit!;
         },
         act: (cubit) async => cubit.loadData(),
         verify: (_) => verify(repository.fetchData()).called(2),
@@ -59,8 +59,8 @@ void main() {
         when(repository.fetchData()).thenAnswer(
           (_) => Future.value(const [RocketVehicle(id: '1')]),
         );
-        await cubit.loadData();
-        expect(cubit.getVehicle('1'), RocketVehicle(id: '1'));
+        await cubit!.loadData();
+        expect(cubit!.getVehicle('1'), RocketVehicle(id: '1'));
       });
     });
   });
