@@ -22,9 +22,9 @@ class LaunchesCubit
     }
   }
 
-  Launch getLaunch(String id) {
+  Launch? getLaunch(String id) {
     if (state.status == RequestStatus.loaded) {
-      return LaunchUtils.getAllLaunches(state.value)
+      return LaunchUtils.getAllLaunches(state.value!)!
           .where((l) => l.id == id)
           .single;
     } else {
