@@ -4,7 +4,7 @@ import '../models/index.dart';
 class LaunchUtils {
   /// Returs the most upcoming launch inside a list, if the list
   /// is sorted by date.
-  static Launch getUpcomingLaunch(List<List<Launch>> launches) {
+  static Launch? getUpcomingLaunch(List<List<Launch>> launches) {
     if (launches != null) {
       return launches[0].first;
     } else {
@@ -14,7 +14,7 @@ class LaunchUtils {
 
   /// Returs the most latest launch inside a list, if the list
   /// is sorted by date.
-  static Launch getLatestLaunch(List<List<Launch>> launches) {
+  static Launch? getLatestLaunch(List<List<Launch>> launches) {
     if (launches != null) {
       return launches[1].first;
     } else {
@@ -23,7 +23,7 @@ class LaunchUtils {
   }
 
   /// Returns all launches combined into one single list.
-  static List<Launch> getAllLaunches(List<List<Launch>> launches) {
+  static List<Launch>? getAllLaunches(List<List<Launch>> launches) {
     if (launches != null) {
       return List.from([...launches[0], ...launches[1]]);
     } else {
