@@ -4,7 +4,7 @@ import 'package:flutter_i18n/flutter_i18n.dart';
 extension Translate on BuildContext {
   String translate(
     final String key, {
-    final Map<String, String> parameters,
+    final Map<String, String>? parameters,
   }) {
     try {
       return FlutterI18n.translate(
